+++ environments-and-requirements/requirements-base.txt
@@ -16,6 +16,7 @@ huggingface-hub
 imageio
 imageio-ffmpeg
 kornia
+npyscreen
 numpy==1.23.*
 omegaconf
 opencv-python
